refactor(pm): move collect_matching_versions into the clean-cache service
collect_matching_versions is cache-clean business logic — its only
consumer is service/clean_cache.rs — but it lived in util/cache.rs with
a pub surface. Move it there as a private fn (typed on CacheEntry, tests
included). matches_pattern stays in util/cache.rs: it is a generic glob
matcher genuinely shared by clean_cache and the workspace filter.
