+++ crates/pm/src/service/clean_cache.rs
@@ -4,10 +4,34 @@ use anyhow::{Context, Result};
 use utoo_ruborist::util::{PackageNameStr, parse_package_spec};
 
 use crate::fs;
-use crate::util::cache::{collect_matching_versions, matches_pattern};
+use crate::util::cache::matches_pattern;
 
 /// A cache entry slated for deletion: (package name, version, on-disk path).
 pub type CacheEntry = (String, String, PathBuf);
+
+/// Collect every version directory of `pkg_name` under `path` whose version
+/// matches `version_pattern`, appending `(pkg_name, version, path)` to
+/// `to_delete`.
+async fn collect_matching_versions(
+    path: &std::path::Path,
+    pkg_name: String,
+    version_pattern: &str,
+    to_delete: &mut Vec<CacheEntry>,
+) -> Result<()> {
+    let mut version_entries = fs::read_dir(path).await?;
+    while let Some(version_entry) = version_entries.next_entry().await? {
+        let version = version_entry.file_name();
+        let version_str = version.to_string_lossy();
+        if matches_pattern(&version_str, version_pattern) {
+            to_delete.push((
+                pkg_name.clone(),
+                version_str.to_string(),
+                version_entry.path(),
+            ));
+        }
+    }
+    Ok(())
+}
 
 /// Walk the global cache directory and collect entries matching `pattern`
 /// (a `name[@version]` spec, both parts may contain wildcards).
@@ -75,5 +99,75 @@ pub async fn delete_cache_entries(to_delete: Vec<CacheEntry>) {
         } else {
             tracing::debug!("Deleted {pkg}@{version}");
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use tempfile::TempDir;
+
+    use super::*;
+
+    async fn setup_test_dir() -> Result<TempDir> {
+        let temp_dir = TempDir::new().unwrap();
+        let base_path = temp_dir.path();
+
+        // Create test version directories
+        fs::create_dir_all(base_path.join("1.0.0")).await?;
+        fs::create_dir_all(base_path.join("1.0.1")).await?;
+        fs::create_dir_all(base_path.join("2.0.0")).await?;
+        fs::create_dir_all(base_path.join("beta-1.0.0")).await?;
+
+        Ok(temp_dir)
+    }
+
+    #[tokio::test]
+    async fn test_collect_matching_versions_exact_match() -> Result<()> {
+        let temp_dir = setup_test_dir().await?;
+        let mut to_delete = Vec::new();
+
+        collect_matching_versions(
+            temp_dir.path(),
+            "test-pkg".to_string(),
+            "1.0.0",
+            &mut to_delete,
+        )
+        .await?;
+
+        assert_eq!(to_delete.len(), 1);
+        assert_eq!(to_delete[0].0, "test-pkg");
+        assert_eq!(to_delete[0].1, "1.0.0");
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_collect_matching_versions_wildcard() -> Result<()> {
+        let temp_dir = setup_test_dir().await?;
+        let mut to_delete = Vec::new();
+
+        collect_matching_versions(
+            temp_dir.path(),
+            "test-pkg".to_string(),
+            "1.*",
+            &mut to_delete,
+        )
+        .await?;
+
+        assert_eq!(to_delete.len(), 2);
+        assert!(to_delete.iter().any(|x| x.1 == "1.0.0"));
+        assert!(to_delete.iter().any(|x| x.1 == "1.0.1"));
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_collect_matching_versions_empty_dir() -> Result<()> {
+        let temp_dir = TempDir::new()?;
+        let mut to_delete = Vec::new();
+
+        collect_matching_versions(temp_dir.path(), "test-pkg".to_string(), "*", &mut to_delete)
+            .await?;
+
+        assert_eq!(to_delete.len(), 0);
+        Ok(())
     }
 }
