+++ crates/pm/src/util/cache.rs
@@ -1,7 +1,7 @@
-use crate::fs;
-
 pub use super::user_config::get_cache_dir;
 
+/// Glob-style match supporting `*` (any), `prefix*`, `*suffix`, `a*b`, and
+/// `@scope/*`. Shared by cache-clean version filtering and workspace filters.
 pub fn matches_pattern(text: &str, pattern: &str) -> bool {
     if pattern == "*" {
         return true;
@@ -38,32 +38,9 @@ pub fn matches_pattern(text: &str, pattern: &str) -> bool {
     text == pattern
 }
 
-pub async fn collect_matching_versions(
-    path: &std::path::Path,
-    pkg_name: String,
-    version_pattern: &str,
-    to_delete: &mut Vec<(String, String, std::path::PathBuf)>,
-) -> anyhow::Result<()> {
-    let mut version_entries = fs::read_dir(path).await?;
-    while let Some(version_entry) = version_entries.next_entry().await? {
-        let version = version_entry.file_name();
-        let version_str = version.to_string_lossy();
-        if matches_pattern(&version_str, version_pattern) {
-            to_delete.push((
-                pkg_name.clone(),
-                version_str.to_string(),
-                version_entry.path(),
-            ));
-        }
-    }
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::fs;
-    use tempfile::TempDir;
 
     #[test]
     fn test_matches_pattern_wildcard() {
@@ -122,68 +99,5 @@ mod tests {
         assert!(!matches_pattern("2.0.0", "1.*"));
         assert!(matches_pattern("1.0.0-beta", "1.0.0*"));
         assert!(!matches_pattern("1.0.1", "1.0.0*"));
-    }
-
-    async fn setup_test_dir() -> anyhow::Result<TempDir> {
-        let temp_dir = TempDir::new().unwrap();
-        let base_path = temp_dir.path();
-
-        // Create test directories
-        fs::create_dir_all(base_path.join("1.0.0")).await?;
-        fs::create_dir_all(base_path.join("1.0.1")).await?;
-        fs::create_dir_all(base_path.join("2.0.0")).await?;
-        fs::create_dir_all(base_path.join("beta-1.0.0")).await?;
-
-        Ok(temp_dir)
-    }
-
-    #[tokio::test]
-    async fn test_collect_matching_versions_exact_match() -> anyhow::Result<()> {
-        let temp_dir = setup_test_dir().await?;
-        let mut to_delete = Vec::new();
-
-        collect_matching_versions(
-            temp_dir.path(),
-            "test-pkg".to_string(),
-            "1.0.0",
-            &mut to_delete,
-        )
-        .await?;
-
-        assert_eq!(to_delete.len(), 1);
-        assert_eq!(to_delete[0].0, "test-pkg");
-        assert_eq!(to_delete[0].1, "1.0.0");
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_collect_matching_versions_wildcard() -> anyhow::Result<()> {
-        let temp_dir = setup_test_dir().await?;
-        let mut to_delete = Vec::new();
-
-        collect_matching_versions(
-            temp_dir.path(),
-            "test-pkg".to_string(),
-            "1.*",
-            &mut to_delete,
-        )
-        .await?;
-
-        assert_eq!(to_delete.len(), 2);
-        assert!(to_delete.iter().any(|x| x.1 == "1.0.0"));
-        assert!(to_delete.iter().any(|x| x.1 == "1.0.1"));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_collect_matching_versions_empty_dir() -> anyhow::Result<()> {
-        let temp_dir = TempDir::new()?;
-        let mut to_delete = Vec::new();
-
-        collect_matching_versions(temp_dir.path(), "test-pkg".to_string(), "*", &mut to_delete)
-            .await?;
-
-        assert_eq!(to_delete.len(), 0);
-        Ok(())
     }
 }
